Scaffolded the Profile resource and associated it with the user resource
This creates the Profile model, controller, route, serializer and tests.

Generated a migration to associate a profile with the current user.

Created CURL scripts to test the index, show, update and destroy.

Currently, a user can create multiple profiles which is not correct. To
do, only allow a user to update a profile after it is created.